# api/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Equipment & Service Management API
  description: API for managing hospital equipment inventory, service requests, and location tracking.
  version: 1.0.0
  contact:
    name: Dmytro Dzhuha, Szenasi Levente
servers:
  - url: http://localhost:8080/v1
    description: Local development
...
paths:
...
  /service-requests:
    get:
      tags: [Service Requests]
      summary: List service requests
      operationId: listServiceRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ServiceRequestStatus'
        - name: equipmentId
          in: query
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
      responses:
        '200':
          description: Paginated service request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequestPage'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [Service Requests]
      summary: Create service request
      description: Medical staff submits a new service/repair request for a piece of equipment.
      operationId: createServiceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequestCreate'
      responses:
        '201':
          description: Service request created with status NEW
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Referenced equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /service-requests/{requestId}:
    parameters:
      - $ref: '#/components/parameters/RequestId'

    get:
      tags: [Service Requests]
      summary: Get service request
      operationId: getServiceRequest
      responses:
        '200':
          description: Service request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Service Requests]
      summary: Update service request
      description: Full update of service request fields including status.
      operationId: updateServiceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequestUpdate'
      responses:
        '200':
          description: Updated service request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot edit a closed service request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Service Requests]
      summary: Delete service request
      description: Only NEW requests that have not been assigned can be deleted.
      operationId: deleteServiceRequest
      responses:
        '204':
          description: Service request deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete a request that is already assigned or in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    EquipmentStatus:
      type: string
      enum:
        - ACTIVE
        - IN_SERVICE
        - DECOMMISSIONED
      description: |
        - `ACTIVE` – in normal use
        - `IN_SERVICE` – undergoing maintenance/repair
        - `DECOMMISSIONED` – retired from service
    ServiceRequestStatus:
      type: string
      enum:
        - NEW
        - ASSIGNED
        - IN_PROGRESS
        - CLOSED
      description: |
        - `NEW` – submitted, awaiting assignment
        - `ASSIGNED` – assigned to a technician
        - `IN_PROGRESS` – technician is actively working on it
        - `CLOSED` – resolved and closed
    Priority:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL
    LocationBase:
      type: object
      required: [building, floor, department, room]
      properties:
        building:
          type: string
          maxLength: 100
          example: "Pavilón A"
        floor:
          type: string
          maxLength: 50
          example: "2. poschodie"
        department:
          type: string
          maxLength: 100
          example: "Kardiológia"
        room:
          type: string
          maxLength: 50
          example: "Miestnosť 204"
        description:
          type: string
          maxLength: 500
...
    Location:
      allOf:
        - $ref: '#/components/schemas/LocationBase'
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            equipmentCount:
              type: integer
              readOnly: true
              description: Number of active equipment items currently at this location
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    EquipmentBase:
      type: object
      required: [name, type, inventoryNumber, locationId]
      properties:
        name:
          type: string
          maxLength: 200
          example: "Ultrazvuk Philips EPIQ 7"
        type:
          type: string
          maxLength: 100
          example: "Ultrasonograf"
        inventoryNumber:
          type: string
          maxLength: 50
          description: Unique inventory/asset number
          example: "INV-2024-00421"
        warrantyExpiry:
          type: string
          format: date
        locationId:
          type: string
          format: uuid
          description: Current location ID
        notes:
          type: string
          maxLength: 2000
...
    Equipment:
      allOf:
        - $ref: '#/components/schemas/EquipmentBase'
        - type: object
          required: [id, status, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/EquipmentStatus'
            location:
              $ref: '#/components/schemas/Location'
            openServiceRequestCount:
              type: integer
              readOnly: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    ServiceRequestBase:
      type: object
      required: [equipmentId, title, description, priority]
      properties:
        title:
          type: string
          maxLength: 200
          example: "Porucha displeja – nereaguje na dotyk"
        description:
          type: string
          maxLength: 5000
          example: "Displej prestáva reagovať po 30 minútach prevádzky."
        priority:
          $ref: '#/components/schemas/Priority'
        equipmentId:
          type: string
          format: uuid
    ServiceRequestCreate:
      allOf:
        - $ref: '#/components/schemas/ServiceRequestBase'
      example:
        equipmentId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        title: "Porucha displeja – nereaguje na dotyk"
        description: "Displej prestáva reagovať po 30 minútach prevádzky."
        priority: "HIGH"
    ServiceRequestUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/ServiceRequestStatus'
      example:
        title: "Porucha displeja – nereaguje na dotyk"
        description: "Displej prestáva reagovať po 30 minútach prevádzky."
        priority: "HIGH"
        status: "IN_PROGRESS"
    ServiceRequest:
      allOf:
        - $ref: '#/components/schemas/ServiceRequestBase'
        - type: object
          required: [id, status, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/ServiceRequestStatus'
            equipment:
              $ref: '#/components/schemas/Equipment'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
      example:
        id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        equipmentId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        title: "Porucha displeja – nereaguje na dotyk"
        description: "Displej prestáva reagovať po 30 minútach prevádzky."
        priority: "HIGH"
        status: "IN_PROGRESS"
        createdAt: "2024-03-15T08:30:00Z"
        updatedAt: "2024-03-16T10:00:00Z"
    ServiceRequestPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ServiceRequest'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: EQUIPMENT_NOT_FOUND
        message:
          type: string
          example: Equipment with id '3fa85f64-...' was not found.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string