# internal/inventoryapi/model_equipment.go
/*
 * Hospital Equipment & Service Management API
 *
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package inventoryapi

import (
	"time"
)

type Equipment struct {

	Name string `json:"name"`

	Type string `json:"type"`

	// Unique inventory/asset number
	InventoryNumber string `json:"inventoryNumber"`

	WarrantyExpiry string `json:"warrantyExpiry,omitempty"`

	// Current location ID
	LocationId string `json:"locationId"`

	Notes string `json:"notes,omitempty"`

	Id string `json:"id"`

	Status EquipmentStatus `json:"status"`

	Location Location `json:"location,omitempty"`

	OpenServiceRequestCount int32 `json:"openServiceRequestCount,omitempty"`

	CreatedAt time.Time `json:"createdAt"`

	UpdatedAt time.Time `json:"updatedAt"`
}