# internal/inventoryapi/model_service_request.go
/*
 * Hospital Equipment & Service Management API
 *
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package inventoryapi

import (
	"time"
)

type ServiceRequest struct {

	Title string `json:"title"`

	Description string `json:"description"`

	Priority Priority `json:"priority"`

	EquipmentId string `json:"equipmentId"`

	Id string `json:"id"`

	Status ServiceRequestStatus `json:"status"`

	Equipment Equipment `json:"equipment,omitempty"`

	CreatedAt time.Time `json:"createdAt"`

	UpdatedAt time.Time `json:"updatedAt"`
}