# internal/inventoryapi/model_service_request_create.go
/*
 * Hospital Equipment & Service Management API
 *
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package inventoryapi

type ServiceRequestCreate struct {

	Title string `json:"title"`

	Description string `json:"description"`

	Priority Priority `json:"priority"`

	EquipmentId string `json:"equipmentId"`
}