# internal/inventoryapi/model_service_request_update.go
/*
 * Hospital Equipment & Service Management API
 *
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package inventoryapi

type ServiceRequestUpdate struct {

	Title string `json:"title,omitempty"`

	Description string `json:"description,omitempty"`

	Priority Priority `json:"priority,omitempty"`

	Status ServiceRequestStatus `json:"status,omitempty"`
}